+++ src/inifix/validation.py
@@ -1,11 +1,24 @@
 import re
+import sys
 from collections.abc import Mapping
-from typing import Any
+from enum import Enum, auto
+from typing import Any, Literal
 
 from inifix._more import always_iterable
 
+if sys.version_info >= (3, 11):
+    from typing import assert_never
+else:
+    from typing_extensions import assert_never
+
 _PARAM_NAME_REGEXP = re.compile(r"[-\.\w]+")
 SCALAR_TYPES = (int, float, bool, str)
+
+
+class SectionsMode(Enum):
+    ALLOW = auto()
+    FORBID = auto()
+    REQUIRE = auto()
 
 
 def _uses_invalid_chars(s: str) -> bool:
@@ -44,19 +57,75 @@ def validate_elementary_item(key: Any, value: Any) -> None:
             )
 
 
-def validate_inifile_schema(d: Mapping, /) -> None:
+def validate_inifile_schema(
+    data: Mapping,
+    /,
+    *,
+    sections: Literal["allow", "forbid", "require"] = "allow",
+) -> None:
     """
-    Raise `ValueError` if and only if the argument is not a
-    valid configuration.
-    """
+    Validate structure of a dictionary as an inifix-compliant configuration.
 
-    for k, v in d.items():
+    Parameters
+    ----------
+    data: dict
+      the candidate configuration to be (in)validated.
+
+    sections: 'allow', 'forbid' or 'require' (default: 'allow')
+      use sections='forbid' to invalidate any section found,
+      or sections='require' to invalidate a sectionless structure.
+      Default mode (sections='allow') permits both.
+
+      .. versionadded: 5.1.0
+
+    Raises
+    ------
+    TypeError: for unrecognized values in parameter sections.
+    ValueError: if and only if data does not conform to the expected schema.
+
+    """
+    match sections:
+        case "allow":
+            sections_mode = SectionsMode.ALLOW
+        case "forbid":
+            sections_mode = SectionsMode.FORBID
+        case "require":
+            sections_mode = SectionsMode.REQUIRE
+        case _:
+            raise TypeError(
+                "Unknown value for parameter sections. "
+                f"Got {sections=!r}, expected 'allow', 'forbid' or 'require'"
+            )
+
+    for k, v in data.items():
         if not isinstance(k, str):
             raise ValueError(
-                f"Invalid schema: received key '{k}' with type '{type(k)}', expected a str"
+                f"Invalid schema: received key '{k}' "
+                f"with type '{type(k)}', expected a str"
             )
+
         if isinstance(v, dict):
-            for kk, vv in v.items():
-                validate_elementary_item(kk, vv)
+            match sections_mode:
+                case SectionsMode.ALLOW | SectionsMode.REQUIRE:
+                    for kk, vv in v.items():
+                        validate_elementary_item(kk, vv)
+                case SectionsMode.FORBID:
+                    raise ValueError(
+                        "Invalid schema: sections were explicitly forbidden, "
+                        f"but one was found under key '{k}'"
+                    )
+                case _:  # pragma: no cover
+                    assert_never(sections_mode)
+
         else:
-            validate_elementary_item(k, v)
+            match sections_mode:
+                case SectionsMode.ALLOW | SectionsMode.FORBID:
+                    validate_elementary_item(k, v)
+                case SectionsMode.REQUIRE:
+                    raise ValueError(
+                        "Invalid schema: sections were explicitly required, "
+                        "but the following key/value pair was found outside of "
+                        f"any section: '{k}', {v}"
+                    )
+                case _:  # pragma: no cover
+                    assert_never(sections_mode)